// onbkp-next/src/services/executed-routines/types.ts
export interface ExecutedRoutines {
    id_executed_routine: string
    id_routine: string
    time_to_start: string
    start_routine: string
    end_routine: string
    start_backup: string
    end_backup: string
    start_compression: string
    end_compression: string
    start_send_cloud: string
    end_send_cloud: string
    situation: string
    file_size: string
    deleted: string
    file_name: string
    observation: string
    bucket_key: string
    created_at: string
}